+++ src/backend/InvenTree/InvenTree/exceptions.py
@@ -19,10 +19,10 @@ logger = structlog.get_logger('inventree')
 
 
 def log_error(
-    path,
-    error_name=None,
-    error_info=None,
-    error_data=None,
+    path: Optional[str] = None,
+    error_name: Optional[str] = None,
+    error_info: Optional[str] = None,
+    error_data: Optional[str] = None,
     scope: Optional[str] = None,
     plugin: Optional[str] = None,
 ):