+++ src/backend/InvenTree/plugin/helpers.py
@@ -108,9 +108,6 @@ def handle_error(error, do_raise: bool = True, do_log: bool = True, log_name: st
 
 def get_entrypoints():
     """Returns list for entrypoints for InvenTree plugins."""
-    # on python before 3.12, we need to use importlib_metadata
-    if sys.version_info < (3, 12):
-        return entry_points().get('inventree_plugins', [])
     return entry_points(group='inventree_plugins')
 
 
@@ -190,13 +187,11 @@ def get_modules(pkg, path=None):
             continue
 
         try:
-            if sys.version_info < (3, 12):
-                module = finder.find_module(name).load_module(name)
-            else:
-                spec = finder.find_spec(name)
-                module = module_from_spec(spec)
-                sys.modules[name] = module
-                spec.loader.exec_module(module)
+            spec = finder.find_spec(name, path)
+            module = module_from_spec(spec)
+            sys.modules[name] = module
+            spec.loader.exec_module(module)
+
             pkg_names = getattr(module, '__all__', None)
             for k, v in vars(module).items():
                 if not k.startswith('_') and (pkg_names is None or k in pkg_names):